use float64 for money amount instead of int

+++ api/requests.go
@@ -4,13 +4,13 @@ type CodeRequest struct {
 	TransactionCode string `json:"transaction_code"`
 }
 type BanknoteCheckRequest struct {
-	TransactionCode string `json:"transaction_code"`
-	Currency        string `json:"currency"`
-	Amount          int    `json:"amount"`
+	TransactionCode string  `json:"transaction_code"`
+	Currency        string  `json:"currency"`
+	Amount          float64 `json:"amount"`
 }
 type FinalizeTransactionRequest struct {
-	UserId    string `json:"user_id"`
-	ATMSecret string `json:"atm_secret"`
-	Currency  string `json:"currency"`
-	Amount    int    `json:"amount"` // negative value means withdrawal
+	UserId    string  `json:"user_id"`
+	ATMSecret string  `json:"atm_secret"`
+	Currency  string  `json:"currency"`
+	Amount    float64 `json:"amount"` // negative value means withdrawal
 }
